+++ src/output/p_aac.cpp
@@ -228,8 +228,8 @@ aac_packetizer_c::process(packet_cptr packet) {
 
 void
 aac_packetizer_c::dump_debug_info() {
-  mxdebug("aac_packetizer_c: queue: %d; buffer size: %d\n",
-          packet_queue.size(), byte_buffer.get_size());
+  mxdebug("aac_packetizer_c: queue: %u; buffer size: %d\n",
+          (unsigned int)packet_queue.size(), byte_buffer.get_size());
 }
 
 connection_result_e
