+++ src/output/p_wavpack.cpp
@@ -76,7 +76,8 @@ wavpack_packetizer_c::process(packet_cptr packet) {
 
 void
 wavpack_packetizer_c::dump_debug_info() {
-  mxdebug("wavpack_packetizer_c: queue: %d\n", packet_queue.size());
+  mxdebug("wavpack_packetizer_c: queue: %u\n",
+          (unsigned int)packet_queue.size());
 }
 
 connection_result_e
